Update radial.c
The formulas for the UTA width and shift for the case of electric and magnetic multipoles are taken from Bar-Shalom et al. 1995. These expressions are wrong for magnetic transitions and for electric transitions of even multipole order. The corrected formulas were implemented (M. Krief & A. Feigel, work in progress).

The various two-electron variance formulas are implemented in:
ConfigEnergyVarianceParts0
p=0 is Q1
p=1 is Q2
p=2 is Q3
p=3 is Q4
p=4 is Q5
p=5 is Q6
p=6 is P1
p=7 is P2
p=8 is P3
p=9 is P4

ConfigEnergyVarianceParts1 
p=0 is P5 
p=1 is P6
p=2 is P7.

The corrections to the formulas are:
For P5,P7,Q2 and the energy shift, the phase (-1)^(ja+jb+1) was replaced by  (-1)^(ja+jb+m2).
For Q3 the phase (-1)^(k'+m2) multiply the two 6js term.
For Q6 the phase (-1)^(ja+jb) was replaced by (-1)^(ja+jb+m2+k') and also two new phases (-1)^k' and (-1)^m2 appear.

For the energy shift,
in the function ConfigEnergyShift
a phase (-1)^m2 was added to the FKT sum.